+++ Insercao.h
@@ -1,7 +1,7 @@
 #ifndef INSERCAO_H
 #define INSERCAO_H
 
-void insercao(int *lista, int inicio, int fim, int *numeroComparacoes, int *numeroTrocas);
+void insercao(int *lista, int inicio, int fim, unsigned long int *numeroComparacoes, unsigned long int *numeroTrocas);
 
 #endif
